refactor(tests): share the ship suites' oracle helpers
Four helpers were written out again per suite as the milestones landed. Measured
with `diff` over the extracted function bodies:

* `oracle(script, args)` — `ship-macos` vs `ship-windows`, 11/11 lines identical,
  doc comment included;
* `oracleExpectingFailure` — 9/9 identical (the doc drifted: macOS carries an
  8-line rationale, Windows compressed it to one line);
* `shipExpectingFailure` — 8/8 identical in `ship-macos`/`ship-windows`, and two
  more near-copies in `ship-msi` and `ship-layout` that differ only by an
  optional `env` argument and a missing `return`;
* `sha256` — one line, `sort -u` over the three copies returns one;
* `ORACLE` — the same `join(MONOREPO_ROOT, '.github', 'ship-oracle')` three times.

An oracle helper is the worst thing to hold two copies of: a suite whose runner
quietly stops discriminating still passes. All five now live in
`tests/e2e/ship/fixture.mjs`, which all four suites already imported.

`ship-layout`'s copy also carried an ORPHANED doc block — the paragraph
describing `shipExpectingFailure` sat above `runCliCapture`'s own doc, two
functions from the one it documented. Moving the function took its doc with it.

`ship-msi`'s `oracle`/`oracleFrom`/`oracleExpectingFailure` are NOT merged: they
hardcode `verify-msi.sh` and one takes a `cwd`, so they are a specialisation
rather than a copy.

Also corrects a count these two byte-identical tests disagreed about: both assert
that a bare `gjsify ship` on Linux still defaults to deb and rpm, and the macOS
comment said "four artifacts" where the Windows one said SIX. Six is right —
`grep -c "finishOn: 'any'"` over `formats.ts` is 6 (deb, rpm, macos-app,
macos-app-zip, windows-dir, windows-dir-zip).

Pure move. `ship`, `ship-layout`, `ship-macos`, `ship-windows`, `ship-signing`,
`ship-from-stage` and `ship-declaration` run 190 tests, all green, the same
assertions before and after. `ship-msi` cannot run here (no `msitools` on this
workstation, and the suite refuses rather than skipping) — it produces the same
21 cancelled tests and the same single `wixl is not on PATH` error as before this
commit, so module load and every import still resolve.

One more stale comment in `ship-layout`, in the file this commit already
touches. It said `readBinaryArch` "returns `null` for PE by design (`payload.ts`
says so), so a Windows payload whose only native file is a `.dll` cannot trip
this check at all", and pointed at `status/open-todos.md` item 5. All four
halves are now false: `payload.ts` implements the PE branch, `payload.spec.ts`
covers the three PE machines, `tests/e2e/ship-windows` refuses an arm64 closure
labelled x64 by name, and the open-todos entry is struck through as CLOSED
(#1354 M3). The ledger was updated and the comment was not.
